Moved files from cmd/retester to pkg/retester
Changed some private functions to public.
Added comments to newly public functions.
