+++ pkg/retester/testdata/testconfig/openshift-config.yaml
